Добавлен point презентер
