+++ markup/add-new-point.html
@@ -77,7 +77,7 @@
 
             <div class="trip-sort__item  trip-sort__item--event">
               <input id="sort-event" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-event" disabled>
-              <label class="trip-sort__btn" for="sort-event">Event</label>
+              <label class="trip-sort__btn" for="sort-event">Point</label>
             </div>
 
             <div class="trip-sort__item  trip-sort__item--time">
@@ -103,13 +103,13 @@
                   <div class="event__type-wrapper">
                     <label class="event__type  event__type-btn" for="event-type-toggle-1">
                       <span class="visually-hidden">Choose event type</span>
-                      <img class="event__type-icon" width="17" height="17" src="img/icons/flight.png" alt="Event type icon">
+                      <img class="event__type-icon" width="17" height="17" src="img/icons/flight.png" alt="Point type icon">
                     </label>
                     <input class="event__type-toggle  visually-hidden" id="event-type-toggle-1" type="checkbox">
 
                     <div class="event__type-list">
                       <fieldset class="event__type-group">
-                        <legend class="visually-hidden">Event type</legend>
+                        <legend class="visually-hidden">Point type</legend>
 
                         <div class="event__type-item">
                           <input id="event-type-taxi-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="taxi">
@@ -253,11 +253,11 @@
 
                     <div class="event__photos-container">
                       <div class="event__photos-tape">
-                        <img class="event__photo" src="img/photos/1.jpg" alt="Event photo">
-                        <img class="event__photo" src="img/photos/2.jpg" alt="Event photo">
-                        <img class="event__photo" src="img/photos/3.jpg" alt="Event photo">
-                        <img class="event__photo" src="img/photos/4.jpg" alt="Event photo">
-                        <img class="event__photo" src="img/photos/5.jpg" alt="Event photo">
+                        <img class="event__photo" src="img/photos/1.jpg" alt="Point photo">
+                        <img class="event__photo" src="img/photos/2.jpg" alt="Point photo">
+                        <img class="event__photo" src="img/photos/3.jpg" alt="Point photo">
+                        <img class="event__photo" src="img/photos/4.jpg" alt="Point photo">
+                        <img class="event__photo" src="img/photos/5.jpg" alt="Point photo">
                       </div>
                     </div>
                   </section>
@@ -269,7 +269,7 @@
               <div class="event">
                 <time class="event__date" datetime="2019-03-18">MAR 18</time>
                 <div class="event__type">
-                  <img class="event__type-icon" width="42" height="42" src="img/icons/taxi.png" alt="Event type icon">
+                  <img class="event__type-icon" width="42" height="42" src="img/icons/taxi.png" alt="Point type icon">
                 </div>
                 <h3 class="event__title">Taxi Amsterdam</h3>
                 <div class="event__schedule">
@@ -307,7 +307,7 @@
               <div class="event">
                 <time class="event__date" datetime="2019-03-18">MAR 18</time>
                 <div class="event__type">
-                  <img class="event__type-icon" width="42" height="42" src="img/icons/flight.png" alt="Event type icon">
+                  <img class="event__type-icon" width="42" height="42" src="img/icons/flight.png" alt="Point type icon">
                 </div>
                 <h3 class="event__title">Flight Chamonix</h3>
                 <div class="event__schedule">
@@ -350,7 +350,7 @@
               <div class="event">
                 <time class="event__date" datetime="2019-03-18">MAR 18</time>
                 <div class="event__type">
-                  <img class="event__type-icon" width="42" height="42" src="img/icons/drive.png" alt="Event type icon">
+                  <img class="event__type-icon" width="42" height="42" src="img/icons/drive.png" alt="Point type icon">
                 </div>
                 <h3 class="event__title">Drive Chamonix</h3>
                 <div class="event__schedule">
@@ -388,7 +388,7 @@
               <div class="event">
                 <time class="event__date" datetime="2019-03-18">MAR 18</time>
                 <div class="event__type">
-                  <img class="event__type-icon" width="42" height="42" src="img/icons/check-in.png" alt="Event type icon">
+                  <img class="event__type-icon" width="42" height="42" src="img/icons/check-in.png" alt="Point type icon">
                 </div>
                 <h3 class="event__title">Check-in Chamonix</h3>
                 <div class="event__schedule">
@@ -426,7 +426,7 @@
               <div class="event">
                 <time class="event__date" datetime="2019-03-19">MAR 19</time>
                 <div class="event__type">
-                  <img class="event__type-icon" width="42" height="42" src="img/icons/sightseeing.png" alt="Event type icon">
+                  <img class="event__type-icon" width="42" height="42" src="img/icons/sightseeing.png" alt="Point type icon">
                 </div>
                 <h3 class="event__title">Sightseeing Chamonix</h3>
                 <div class="event__schedule">
@@ -469,7 +469,7 @@
               <div class="event">
                 <time class="event__date" datetime="2019-03-19">MAR 19</time>
                 <div class="event__type">
-                  <img class="event__type-icon" width="42" height="42" src="img/icons/drive.png" alt="Event type icon">
+                  <img class="event__type-icon" width="42" height="42" src="img/icons/drive.png" alt="Point type icon">
                 </div>
                 <h3 class="event__title">Drive Geneva</h3>
                 <div class="event__schedule">
@@ -499,7 +499,7 @@
               <div class="event">
                 <time class="event__date" datetime="2019-03-19">MAR 19</time>
                 <div class="event__type">
-                  <img class="event__type-icon" width="42" height="42" src="img/icons/flight.png" alt="Event type icon">
+                  <img class="event__type-icon" width="42" height="42" src="img/icons/flight.png" alt="Point type icon">
                 </div>
                 <h3 class="event__title">Flight Geneva</h3>
                 <div class="event__schedule">
@@ -542,7 +542,7 @@
               <div class="event">
                 <time class="event__date" datetime="2019-03-20">MAR 20</time>
                 <div class="event__type">
-                  <img class="event__type-icon" width="42" height="42" src="img/icons/drive.png" alt="Event type icon">
+                  <img class="event__type-icon" width="42" height="42" src="img/icons/drive.png" alt="Point type icon">
                 </div>
                 <h3 class="event__title">Drive Geneva</h3>
                 <div class="event__schedule">
@@ -572,7 +572,7 @@
               <div class="event">
                 <time class="event__date" datetime="2019-03-20">MAR 20</time>
                 <div class="event__type">
-                  <img class="event__type-icon" width="42" height="42" src="img/icons/sightseeing.png" alt="Event type icon">
+                  <img class="event__type-icon" width="42" height="42" src="img/icons/sightseeing.png" alt="Point type icon">
                 </div>
                 <h3 class="event__title">Sightseeing Geneva</h3>
                 <div class="event__schedule">
