+++ markup/list-empty.html
@@ -57,11 +57,11 @@
         <section class="trip-events">
           <h2 class="visually-hidden">Trip events</h2>
 
-          <p class="trip-events__msg">Click New Event to create your first point</p>
+          <p class="trip-events__msg">Click New Point to create your first point</p>
 
           <!--
             Значение отображаемого текста зависит от выбранного фильтра:
-              * Everthing – 'Click New Event to create your first point'
+              * Everthing – 'Click New Point to create your first point'
               * Past — 'There are no past events now';
               * Future — 'There are no future events now'.
           -->
