+++ markup/list-filter.html
@@ -77,7 +77,7 @@
 
             <div class="trip-sort__item  trip-sort__item--event">
               <input id="sort-event" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-event" disabled>
-              <label class="trip-sort__btn" for="sort-event">Event</label>
+              <label class="trip-sort__btn" for="sort-event">Point</label>
             </div>
 
             <div class="trip-sort__item  trip-sort__item--time">
@@ -101,7 +101,7 @@
               <div class="event">
                 <time class="event__date" datetime="2019-03-18">MAR 18</time>
                 <div class="event__type">
-                  <img class="event__type-icon" width="42" height="42" src="img/icons/taxi.png" alt="Event type icon">
+                  <img class="event__type-icon" width="42" height="42" src="img/icons/taxi.png" alt="Point type icon">
                 </div>
                 <h3 class="event__title">Taxi Amsterdam</h3>
                 <div class="event__schedule">
@@ -139,7 +139,7 @@
               <div class="event">
                 <time class="event__date" datetime="2019-03-18">MAR 18</time>
                 <div class="event__type">
-                  <img class="event__type-icon" width="42" height="42" src="img/icons/flight.png" alt="Event type icon">
+                  <img class="event__type-icon" width="42" height="42" src="img/icons/flight.png" alt="Point type icon">
                 </div>
                 <h3 class="event__title">Flight Chamonix</h3>
                 <div class="event__schedule">
@@ -182,7 +182,7 @@
               <div class="event">
                 <time class="event__date" datetime="2019-03-18">MAR 18</time>
                 <div class="event__type">
-                  <img class="event__type-icon" width="42" height="42" src="img/icons/drive.png" alt="Event type icon">
+                  <img class="event__type-icon" width="42" height="42" src="img/icons/drive.png" alt="Point type icon">
                 </div>
                 <h3 class="event__title">Drive Chamonix</h3>
                 <div class="event__schedule">
@@ -220,7 +220,7 @@
               <div class="event">
                 <time class="event__date" datetime="2019-03-18">MAR 18</time>
                 <div class="event__type">
-                  <img class="event__type-icon" width="42" height="42" src="img/icons/check-in.png" alt="Event type icon">
+                  <img class="event__type-icon" width="42" height="42" src="img/icons/check-in.png" alt="Point type icon">
                 </div>
                 <h3 class="event__title">Check-in Chamonix</h3>
                 <div class="event__schedule">
