+++ src/view/point-add.js
